Safari-611.1.10/LayoutTests/scrollingcoordinator/mac/latching/main-frame-back-swipe-expected.txt

Swipes should rubber-band on each side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Swipe to the left
begin swipe
PASS minXOffset is 0
PASS maxXOffset is -1000

Swipe to the right
PASS minXOffset is 0
PASS maxXOffset > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
